Removed unused tookit functions and switched to using subprocess
