+++ GutFunFind/detect/hmmer_search/__init__.py
@@ -1,11 +1,12 @@
 import os
 from configparser import ConfigParser
 from typing import IO, List, Union
+import subprocess
 
 from Bio.SearchIO._model.query import QueryResult
 from Bio import SearchIO
 
-from GutFunFind.toolkit.base import *
+from GutFunFind.toolkit.utility import *
 from GutFunFind.detect.hmmer_search import *
 from GutFunFind.detect.hmmer_search.hmmer_filter import hmmer_filter
 
@@ -42,9 +43,8 @@ def pipeline(config: dict,
         cmd.insert(2,config["hmmer"]["hmmer.threads"])
 
     # 3 run command line
-    # check the protein_file+".blast.xml" existence
-    res = execute(cmd)
-    if res.return_code:
+    res = subprocess.run(cmd)
+    if res.returncode != 0:
         raise RuntimeError("Failed to run: {}".format(" ".join(cmd)))
 
     # 4 read the output from previous step
